cleanup: rename internal/commands to internal/cmd
