+++ internal/cmd/debug/debug.go
