+++ internal/cmd/expand/expand.go
