+++ internal/cmd/expand/expand_test.go
