+++ internal/cmd/inits/data/aliases.sh
