+++ internal/cmd/inits/data/git_wrapper.fish
