+++ internal/cmd/inits/data/git_wrapper.sh
