+++ internal/cmd/inits/data/status_shortcuts.fish
