+++ internal/cmd/inits/data/status_shortcuts.sh
