+++ internal/cmd/inits/scripts.go
