+++ internal/cmd/status/color.go
