+++ internal/cmd/status/render.go
