+++ internal/cmd/status/render_test.go
