+++ internal/cmd/status/status.go
