+++ internal/cmd/status/testdata/.gitattributes
