+++ internal/cmd/status/testdata/statuslist-complex_mix.display.ansi.golden
