+++ internal/cmd/status/testdata/statuslist-empty.display.ansi.golden
