+++ internal/cmd/status/testdata/statuslist-longlist.display.ansi.golden
