+++ internal/cmd/status/testdata/statuslist-longlist.parsedata.txt.golden
