+++ internal/cmd/status/testdata/statuslist-subdirectory.display.ansi.golden
