+++ internal/cmd/status/testdata/statuslist-subdirectory.parsedata.txt.golden
