+++ internal/cmd/status/testdata/statuslist-type_changes.display.ansi.golden
