+++ internal/cmd/status/testdata/statuslist-type_changes.parsedata.txt.golden
