+++ internal/cmd/status/testdata/statuslist-with_branch_ahead.display.ansi.golden
