+++ internal/cmd/status/testdata/statuslist-with_branch_ahead.parsedata.txt.golden
