+++ internal/cmd/status/testdata/statuslist-with_staged_files.display.ansi.golden
