+++ internal/cmd/status/testdata/statuslist-with_staged_files.parsedata.txt.golden
